+++ docs/tr2/CHANGELOG.md
@@ -55,6 +55,7 @@
 - changed music triggers that match the level's default ambient track to automatically be treated as ambient if omitted from `ambient_tracks` (#4181)
 - changed the Pause key to no longer work when Lara's dead (similar to TR1)
 - changed the `-q`/`--quiet` argument to no longer silence warnings
+- changed the bars options to be placed in its own tab
 - changed the `Remember Guns between Levels` option to also apply to whether or not Lara starts with those guns equipped
 - changed sprites to respect the water tint if placed underwater
 - changed the FOV formula to be consistent between games